requirements.txt
pandas
requests
python-decouple
sqlalchemy